pyproject.toml
[project]
name = "fansite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
